quite a bit of stuff.
Added functions to get the rml of various elements to facilitate editing.
Added a way to set a fake path when loading a document from memory to help find rcss files and stuff.
Added a function to context to get an element at a given point.
Added a zoom level that scales the content of a rml context based on a user defined zoom level.
Added a unit gsp, that only scales with the global scale factor. Good for scrollbars and things that should not change with context zoom.
Added a way to clear the css properties of an element.
Made it so the select controls properly recreate their input rml.
Made it so you can select the contents of a text box by pressing ctrl + a.
Added a way to set the color of a decorator, allows coloring of textures.
Made it so templates can be configured to either append or prepend their head, useful when using multiple templates.
Made it so element style local properties can be cleared.
Added a JoinPath function in the SystemInterface to join two paths together in a potentially custom way.
Fixed the way the style sheet cache determines if a file is in the cache, using full paths now just the filenames was causing collisions.